[Disk Manager] issue-3419: block on external filesystem create/delete task

+++ cloud/disk_manager/internal/pkg/services/filesystem/create_filesystem_task.go
@@ -11,15 +11,17 @@ import (
 	"github.com/ydb-platform/nbs/cloud/disk_manager/internal/pkg/services/filesystem/protos"
 	"github.com/ydb-platform/nbs/cloud/tasks"
 	"github.com/ydb-platform/nbs/cloud/tasks/errors"
+	"github.com/ydb-platform/nbs/cloud/tasks/headers"
 )
 
 ////////////////////////////////////////////////////////////////////////////////
 
 type createFilesystemTask struct {
-	storage resources.Storage
-	factory nfs.Factory
-	request *protos.CreateFilesystemRequest
-	state   *protos.CreateFilesystemTaskState
+	storage   resources.Storage
+	factory   nfs.Factory
+	scheduler tasks.Scheduler
+	request   *protos.CreateFilesystemRequest
+	state     *protos.CreateFilesystemTaskState
 }
 
 func (t *createFilesystemTask) Save() ([]byte, error) {
@@ -41,12 +43,6 @@ func (t *createFilesystemTask) Run(
 	ctx context.Context,
 	execCtx tasks.ExecutionContext,
 ) error {
-
-	client, err := t.factory.NewClient(ctx, t.request.Filesystem.ZoneId)
-	if err != nil {
-		return err
-	}
-	defer client.Close()
 
 	selfTaskID := execCtx.GetTaskID()
 
@@ -75,15 +71,40 @@ func (t *createFilesystemTask) Run(
 		)
 	}
 
-	err = client.Create(ctx, t.request.Filesystem.FilesystemId, nfs.CreateFilesystemParams{
-		CloudID:     t.request.CloudId,
-		FolderID:    t.request.FolderId,
-		BlocksCount: t.request.BlocksCount,
-		BlockSize:   t.request.BlockSize,
-		Kind:        t.request.Kind,
-	})
-	if err != nil {
-		return err
+	if filesystemMeta.Kind == "external" {
+		taskID, err := t.scheduler.ScheduleTask(
+			headers.SetIncomingIdempotencyKey(ctx, selfTaskID+"_run"),
+			"filesystem.CreateExternalFilesystem",
+			"",
+			t.request,
+		)
+		if err != nil {
+			return err
+		}
+
+		t.state.CreateExternalFilesystemTaskID = taskID
+
+		_, err = t.scheduler.WaitTask(ctx, execCtx, taskID)
+		if err != nil {
+			return err
+		}
+	} else {
+		client, err := t.factory.NewClient(ctx, t.request.Filesystem.ZoneId)
+		if err != nil {
+			return err
+		}
+		defer client.Close()
+
+		err = client.Create(ctx, t.request.Filesystem.FilesystemId, nfs.CreateFilesystemParams{
+			CloudID:     t.request.CloudId,
+			FolderID:    t.request.FolderId,
+			BlocksCount: t.request.BlocksCount,
+			BlockSize:   t.request.BlockSize,
+			Kind:        t.request.Kind,
+		})
+		if err != nil {
+			return err
+		}
 	}
 
 	filesystemMeta.CreatedAt = time.Now()
@@ -94,12 +115,6 @@ func (t *createFilesystemTask) Cancel(
 	ctx context.Context,
 	execCtx tasks.ExecutionContext,
 ) error {
-
-	client, err := t.factory.NewClient(ctx, t.request.Filesystem.ZoneId)
-	if err != nil {
-		return err
-	}
-	defer client.Close()
 
 	selfTaskID := execCtx.GetTaskID()
 
@@ -120,9 +135,34 @@ func (t *createFilesystemTask) Cancel(
 		)
 	}
 
-	err = client.Delete(ctx, t.request.Filesystem.FilesystemId)
-	if err != nil {
-		return err
+	if fs.Kind == "external" {
+		taskID, err := t.scheduler.ScheduleTask(
+			headers.SetIncomingIdempotencyKey(ctx, selfTaskID+"_run"),
+			"filesystem.CreateExternalFilesystem",
+			"",
+			t.request,
+		)
+		if err != nil {
+			return err
+		}
+
+		t.state.CreateExternalFilesystemTaskID = taskID
+
+		_, err = t.scheduler.WaitTask(ctx, execCtx, taskID)
+		if err != nil {
+			return err
+		}
+	} else {
+		client, err := t.factory.NewClient(ctx, t.request.Filesystem.ZoneId)
+		if err != nil {
+			return err
+		}
+		defer client.Close()
+
+		err = client.Delete(ctx, t.request.Filesystem.FilesystemId)
+		if err != nil {
+			return err
+		}
 	}
 
 	return t.storage.FilesystemDeleted(
